chore(sweep): move remaining black-box tests to crate-level integration tests
Final pass of the test reorganisation. The pattern is now consistently
applied: black-box tests live in `<crate>/tests/`; inline `mod tests`
is reserved for tests that need module-private access.

Moved to `<crate>/tests/`:

- `jar-kernel/src/{genesis, kernel_assist, kernel, state}.rs` →
  `jar-kernel/tests/`. The kernel_assist test needs the
  `javm::kernel_assist::KernelAssist` trait in scope (its methods
  are private inherent fns; the trait gives the public surface).
- `javm-transpiler/src/{emitter, layout, program}.rs` →
  `javm-transpiler/tests/`. The remaining three (`assembler`, `lib`,
  `riscv`) stay inline — they exercise private fields (`labels`),
  private helpers (`parse_signed_imm`, `map_register`, `decode_*_imm`),
  and the like.
- `nub-host-common/src/log_level.rs` → `nub-host-common/tests/`.

Audited and kept inline (genuinely need private access):

- `javm/src/{callstack, ecall, frame, image_cache, kernel_assist, vm}.rs`
  — tests rely on `use super::*;` to pull in transitively re-exported
  trait/type names that aren't reachable from outside via plain
  imports.
- `javm-recompiler-x86/src/{asm, predecode}.rs` — tests use
  `#[cfg(test)]`-only helpers (`Assembler::code_bytes`,
  `compute_skip`, etc.) that integration tests can't see.
- `nub-host-common/src/arch/amd64/vmem.rs` — `pub(in crate::vmem)`
  helpers.
- `nub-host-kvm/src/**` (10 files) — almost entirely `pub(crate)`
  modules (`hypervisor::regs`, `hypervisor::virtual_machine`,
  `sandbox::outb`) plus private fields on `SandboxMemoryLayout`
  (`code_size`, `heap_size`, `BASE_ADDRESS`) and `Assembler`-style
  private state.

End state: 25 inline `mod tests` blocks moved out; 23 stay inline
because each genuinely needs module-private access. Per-crate test
counts unchanged.

Verified: workspace build/clippy/fmt/test/doc green (505 tests);
nub smoke green on both Local + Hyperlight backends.
