+++ rust/javm-transpiler/src/emitter.rs
@@ -17,15 +17,3 @@ pub fn pack_bitmask(bitmask: &[u8]) -> Vec<u8> {
     }
     packed
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_pack_bitmask() {
-        assert_eq!(pack_bitmask(&[1, 1, 1]), vec![0x07]);
-        assert_eq!(pack_bitmask(&[1, 0, 1, 0, 1, 0, 1, 0]), vec![0x55]);
-        assert_eq!(pack_bitmask(&[1, 0, 1, 0, 1, 0, 1, 0, 1]), vec![0x55, 0x01]);
-    }
-}
