+++ rust/javm-transpiler/src/layout.rs
@@ -136,32 +136,3 @@ impl ProgramLayout {
         self.data_caps().map(|d| d.page_count).sum()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn layout_minimal_stack_only() {
-        let l = ProgramLayout::compute(1, 0, 0, 0);
-        assert_eq!(l.stack.cap_index, STACK_CAP_INDEX);
-        assert_eq!(l.stack.base_page, 0);
-        assert_eq!(l.stack.page_count, 1);
-        assert!(l.ro.is_none());
-        assert!(l.rw.is_none());
-        assert!(l.heap.is_none());
-        assert_eq!(l.stack_top(), 4096);
-        assert_eq!(l.total_data_pages(), 1);
-    }
-
-    #[test]
-    fn layout_full_stack_ro_rw_heap() {
-        let l = ProgramLayout::compute(2, 1, 1, 4);
-        assert_eq!(l.stack.base_page, 0);
-        assert_eq!(l.ro.as_ref().unwrap().base_page, 2);
-        assert_eq!(l.rw.as_ref().unwrap().base_page, 3);
-        assert_eq!(l.heap.as_ref().unwrap().base_page, 4);
-        assert_eq!(l.stack_top(), 2 * 4096);
-        assert_eq!(l.total_data_pages(), 2 + 1 + 1 + 4);
-    }
-}
